# docs/swagger/collections.yaml
openapi: 3.0.0
servers:
  - description: Listmonk Developement Server
    url: http://localhost:9000/api
info:
  version: "1.0.0"
  title: Listmonk
  description: The API collection for listmonk
  license:
    name: AGPL-3.0 license
    url: https://github.com/knadh/listmonk/blob/master/LICENSE
security:
  - bearerAuth: []
  - basicAuth: []
  - apiTokenAuth: []
...
paths:
...
  /templates:
    get:
      description: handles retrieval of templates
      operationId: getTemplates
      tags:
        - Templates
      parameters:
        - in: query
          name: no_body
          description: boolean flag for response with/without body
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Template"
    post:
      description: handles template creation
      operationId: createTemplate
      tags:
        - Templates
      requestBody:
        description: new template info
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTemplate"
      responses:
        "200":
          description: response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Template"
...
  /templates/preview:
    post:
      description: get the HTML preview of a template.
      operationId: previewTemplate
      tags:
        - Templates
      requestBody:
        required: true
        description: template parameters
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                template_type:
                  description: type of template
                  type: string
                body:
                  description: template body
                  type: string
      responses:
        "200":
          description: response
          content:
            text/html:
              schema:
                type: string
                example: <p>Hi there</p>
...
  "/templates/{template_id}/clone":
    post:
      tags:
        - Templates
      description: clones a template into a new template.
      operationId: cloneTemplate
      parameters:
        - in: path
          name: template_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateCloneRequest"
      responses:
        "200":
          description: cloned template
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Template"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
    apiTokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Use the header format: `Authorization: token api_user:token`'
  schemas:
...
    Template:
      type: object
      properties:
        id:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
        name:
          type: string
        body:
          type: string
        body_source:
          type: string
        subject:
          type: string
        type:
          type: string
        is_default:
          type: boolean
    NewTemplate:
      type: object
      required:
        - name
        - type
        - body
      properties:
        name:
          type: string
          description: Name of the template
        type:
          type: string
          enum: [campaign, campaign_visual, tx]
          description: Type of the template
        subject:
          type: string
          description: Subject line for the template (only for tx)
        body_source:
          type: string
          description: JSON source for the email-builder template (only for campaign_visual)
        body:
          type: string
          description: HTML body of the template
...
    TemplateCloneRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        subject:
          type: string
...